+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 See the [Migration Guide](guides/migration_guide.md) for the details of breaking changes between versions.
 
+## 1.0.0
+
+### Improvements
+
+- Test coverage 100%.
+
 ## 0.1.2
 
 ### Fixes